# openapi/omnora.v1.yaml
openapi: 3.1.0
info:
  title: Omnora REST API
  version: 0.1.0
  summary: OpenAPI contract for the first Omnora version.
  description: >
    Contract aligned to the runnable Omnora REST surface. Mount content access is a
    pure account whitelist: effective permission is the lower of current space
    membership and the current mount grant. System administrator status controls
    the control plane and never substitutes for a mount grant. Stale starter paths
    are marked x-omnora-stale and should not be used by new clients.
...
servers:
  - url: /api/v1
...
security:
  - cookieAuth: []
...
paths:
...
  /shares:
    get:
      tags: [Shares]
      summary: List shares visible to the caller.
      x-omnora-route-group: rest
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Shares.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Share"
    post:
      tags: [Shares]
      summary: Create a share for a file or folder.
      description: Requires effective manager permission on the target mount and allowPublicShares=true. The returned secret is shown once.
      x-omnora-route-group: rest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareRequest"
      responses:
        "201":
          description: Share created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateShareResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /shares/{shareId}:
    x-omnora-stale: true
    get:
      tags: [Shares]
      summary: Read share metadata.
      x-omnora-route-group: rest
      parameters:
        - $ref: "#/components/parameters/ShareId"
      responses:
        "200":
          description: Share metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Share"
    patch:
      tags: [Shares]
      summary: Update share limits or revoke state.
      x-omnora-route-group: rest
      parameters:
        - $ref: "#/components/parameters/ShareId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShareRequest"
      responses:
        "200":
          description: Share updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Share"
    delete:
      tags: [Shares]
      summary: Revoke a share.
      x-omnora-route-group: rest
      parameters:
        - $ref: "#/components/parameters/ShareId"
      responses:
        "204":
          description: Share revoked.
...
  /admin/shares:
    get:
      tags: [Admin]
      summary: Implemented endpoint GET /admin/shares.
      x-omnora-route-group: rest
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /admin/spaces/{spaceId}/members:
    get:
      tags: [Admin]
      summary: Implemented endpoint GET /admin/spaces/spaceId/members.
      x-omnora-route-group: rest
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AI Token or session-bound API token
    cookieAuth:
      type: apiKey
      in: cookie
      name: omnora_session
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ShareId:
      name: shareId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/OpaqueId"
...
  responses:
...
    Unauthorized:
      description: Authentication failed or credentials are not currently valid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The current authorization formula denies this operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    OpaqueId:
      type: string
      minLength: 8
      maxLength: 128
      description: Opaque, non-host-path identifier.
    Error:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - unauthorized
            - forbidden
            - route_group_disabled
            - not_found
            - invalid_cursor
            - quota_exceeded
            - rate_limited
            - mount_identity_unverifiable
            - mount_conflict
            - mount_offline
            - readonly_mount
            - object_changed
            - preview_not_supported
            - share_password_required
            - share_invalid
            - upload_conflict
        detail:
          type: string
...
    Share:
      type: object
      description: Share metadata. List responses never include the one-time secret or URL fragment.
      required: [id, publicId, spaceId, mountId, relativePath, allowPreview, allowDownload, usedVisits, usedDownloads, expiresAt, status]
      properties:
        id:
          $ref: "#/components/schemas/OpaqueId"
        publicId:
          type: string
        spaceId:
          $ref: "#/components/schemas/OpaqueId"
        spaceName:
          type: string
        mountId:
          $ref: "#/components/schemas/OpaqueId"
        mountName:
          type: string
        relativePath:
          type: string
        creatorEmail:
          type: string
        creatorDisplayName:
          type: string
        status:
          type: string
          enum: [active, expired, revoked]
        allowPreview:
          type: boolean
        allowDownload:
          type: boolean
        maxVisits:
          type: integer
          minimum: 1
        usedVisits:
          type: integer
          minimum: 0
        maxDownloads:
          type: integer
          minimum: 1
        usedDownloads:
          type: integer
          minimum: 0
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
    CreateShareRequest:
      type: object
      additionalProperties: false
      required: [spaceId, mountId, relativePath]
      properties:
        spaceId:
          $ref: "#/components/schemas/OpaqueId"
        mountId:
          $ref: "#/components/schemas/OpaqueId"
        relativePath:
          type: string
        password:
          type: string
          format: password
        expiresAt:
          type: string
          format: date-time
        allowPreview:
          type: boolean
          default: true
        allowDownload:
          type: boolean
          default: true
        maxVisits:
          type: integer
          minimum: 1
        maxDownloads:
          type: integer
          minimum: 1
    CreateShareResponse:
      type: object
      required: [id, publicId, secret, fragment, expiresAt]
      properties:
        id:
          $ref: "#/components/schemas/OpaqueId"
        publicId:
          type: string
        secret:
          type: string
          description: One-time returned secret. The server stores only its hash and never returns it again.
        fragment:
          type: string
          description: One-time `publicId.secret` fragment value for constructing the public share URL.
        expiresAt:
          type: string
          format: date-time
    UpdateShareRequest:
      type: object
      properties:
        expiresAt:
          type: string
          format: date-time
        allowPreview:
          type: boolean
        allowDownload:
          type: boolean
        maxAccesses:
          type: integer
          minimum: 1
        maxDownloads:
          type: integer
          minimum: 1
        revoked:
          type: boolean
...